+++ SSRenderer/SSRenderer/Public/RenderAssetSerializer/RenderAssetSerializeFunctions.h
@@ -17,7 +17,7 @@ struct MeshRawDataBase;
 
 
 
-SSRENDERER_MODULE int32 AppendDataFromMeshAsset(
+SSRENDERER_MODULE int64 AppendDataFromMeshAsset(
 	SS::PooledList<byte>& Data,
 	const MeshRawDataBase* MeshData);
 
@@ -28,15 +28,15 @@ SSRENDERER_MODULE int64 AppendDataFromMdlcAsset(
 	SS::SHasherW MdlcChildNameSpaceToReplace);
 
 
-SSRENDERER_MODULE int32 FillMeshAssetHeaaderOnly(
+SSRENDERER_MODULE int64 FillMeshAssetHeaaderOnly(
 	MeshRawDataDefault*& InOutMeshRawData,
 	const SS::PooledList<byte>& Data,
-	int Offset = 0);
+	int64 Offset = 0);
 
-SSRENDERER_MODULE int32 FillMeshRawDataFromData(
+SSRENDERER_MODULE int64 FillMeshRawDataFromData(
 	MeshRawDataDefault*& InOutMeshRawData,
 	const SS::PooledList<byte>& Data,
-	int Offset = 0);
+	int64 Offset = 0);
 
 
 SSRENDERER_MODULE int64 AppendApakDataFromAssetList(
@@ -54,7 +54,7 @@ SSRENDERER_MODULE int64 CreateAssetsFromApakData(
 	SS::SHasherW AssetNamespace,
 	int Offset = 0);
 
-SSRENDERER_MODULE int32 CreateMeshAssetFromData(
+SSRENDERER_MODULE int64 CreateMeshAssetFromData(
 	IMeshAsset*& OutMeshAsset,
 	SS::SHasherW AssetName,
 	SS::SHasherW AssetPath,
